Updates NixOS from 23.05 to 23.11, updating nixos and xmobar config.
Updates xmobar to use Alsa plugin instead of Volume plugin, as the
volume plugin was not working... also Alsa should be more
efficient/responsive anyways.
Updates the sources for nixos and home-manager sources, updates
home-manager fetchPypi method, and a handful of moved settings in
configuration.nix, including updating the system.stateVersion and
removing the explicitly set nix version, as the default nix version
was increased to what I use in 23.11.