Changes to neovim config, enabling todo-comments, and several new keybindings.
todo-comments highlights todos nicely and allows to jump between them.
new keybindings are from a reddit post, and are for fixing last spelling
change (ctrl-L), adding a semicolon or comma to the end of the line (may
need to disable these, time will tell) and for duplicating a line and
commenting out the original (ycc) (like to change a config value but
still have a recording of the original).